filter only high rate and show contacs with request permission

+++ app/src/main/java/com/seventhstar/films/room/FavoritesDao.kt
@@ -8,7 +8,7 @@ interface FavoritesDao {
     fun all(): List<FavoritesEntity>
 
     @Query("SELECT * FROM FavoritesEntity WHERE filmId LIKE :id")
-    fun getFavoriteByID(id: Long): List<FavoritesEntity>
+    fun getFavoriteByID(id: Int): List<FavoritesEntity>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     fun insert(entity: FavoritesEntity)
